# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khabar News App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
  displayName: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: string;
  username: string;
  email: string;
  displayName: string;
  bio?: string | null;
  avatarUrl?: string | null;
  followersCount: number;
  followingCount: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserProfile = User & {
  isFollowing: boolean;
};

export interface UsersListResponse {
  users: UserProfile[];
}

export interface UpdateProfileRequest {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface Article {
  id: string;
  title: string;
  summary: string;
  content: string;
  imageUrl?: string | null;
  sourceUrl: string;
  sourceName: string;
  category: string;
  publishedAt: string;
  likesCount: number;
  commentsCount: number;
  savesCount: number;
  isLiked: boolean;
  isSaved: boolean;
}

export type ArticleDetail = Article;

export interface NewsFeedResponse {
  articles: Article[];
  total: number;
  page: number;
  hasMore: boolean;
}

export interface LikeResponse {
  liked: boolean;
  likesCount: number;
}

export interface SaveResponse {
  saved: boolean;
  savesCount: number;
}

export interface Comment {
  id: string;
  articleId: string;
  userId: string;
  content: string;
  createdAt: string;
  user: User;
}

export interface CommentsResponse {
  comments: Comment[];
}

export interface AddCommentRequest {
  content: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  like: "like",
  comment: "comment",
  follow: "follow",
  share: "share",
} as const;

export interface Notification {
  id: string;
  type: NotificationType;
  message: string;
  isRead: boolean;
  createdAt: string;
  actor: User;
  articleId?: string | null;
}

export interface NotificationsResponse {
  notifications: Notification[];
  unreadCount: number;
}

export interface FollowResponse {
  following: boolean;
}

export interface Conversation {
  id: string;
  otherUser: User;
  lastMessage?: string | null;
  lastMessageAt?: string | null;
  unreadCount: number;
}

export interface ConversationsResponse {
  conversations: Conversation[];
}

export interface CreateConversationRequest {
  userId: string;
}

export interface DirectMessage {
  id: string;
  conversationId: string;
  senderId: string;
  content: string;
  articleId?: string | null;
  article?: Article | null;
  createdAt: string;
}

export interface MessagesResponse {
  messages: DirectMessage[];
}

export interface SendMessageRequest {
  content: string;
  articleId?: string;
}

export type GetNewsFeedParams = {
  page?: number;
  limit?: number;
  category?: string;
};

export type SearchNewsParams = {
  q: string;
  page?: number;
};

export type SearchUsersParams = {
  q: string;
};