# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Khabar News App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /messages/conversations:
    get:
      operationId: getConversations
      tags: [messages]
      summary: Get conversations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationsResponse"
    post:
      operationId: createOrGetConversation
      tags: [messages]
      summary: Create or get conversation with user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /messages/conversations/{id}:
    get:
      operationId: getConversationMessages
      tags: [messages]
      summary: Get messages in a conversation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesResponse"
  /messages/conversations/{id}/send:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectMessage"
  /ai/explain/{articleId}:
    post:
      operationId: explainArticle
      tags: [ai]
      summary: AI explains article in detail (streaming SSE)
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Streaming SSE response
          content:
            text/event-stream:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        displayName:
          type: string
      required:
        - username
        - email
        - password
        - displayName
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        followersCount:
          type: integer
        followingCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - displayName
        - followersCount
        - followingCount
        - createdAt
...
    Article:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        sourceUrl:
          type: string
        sourceName:
          type: string
        category:
          type: string
        publishedAt:
          type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        savesCount:
          type: integer
        isLiked:
          type: boolean
        isSaved:
          type: boolean
      required:
        - id
        - title
        - summary
        - content
        - sourceUrl
        - sourceName
        - category
        - publishedAt
        - likesCount
        - commentsCount
        - savesCount
        - isLiked
        - isSaved
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        otherUser:
          $ref: "#/components/schemas/User"
        lastMessage:
          type: string
          nullable: true
        lastMessageAt:
          type: string
          nullable: true
        unreadCount:
          type: integer
      required:
        - id
        - otherUser
        - unreadCount
    ConversationsResponse:
      type: object
      properties:
        conversations:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
      required:
        - conversations
    CreateConversationRequest:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
    DirectMessage:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        senderId:
          type: string
        content:
          type: string
        articleId:
          type: string
          nullable: true
        article:
          $ref: "#/components/schemas/Article"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - conversationId
        - senderId
        - content
        - createdAt
    MessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/DirectMessage"
      required:
        - messages
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
        articleId:
          type: string
      required:
        - content