# lib/api-zod/src/generated/types/directMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Khabar News App API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";

export interface DirectMessage {
  id: string;
  conversationId: string;
  senderId: string;
  content: string;
  articleId?: string | null;
  article?: Article | null;
  createdAt: string;
}